bot/requirements.txt
python-telegram-bot==21.4
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-cloud-build==3.25.0
google-api-core==2.19.1
